Доработка кода, редактирование, добавление недостоющего, разбиение методов

+++ app/src/main/java/com/nehvedovich/vladimir/pogoda/screens/rest/entites/CloudsRestModel.java
@@ -3,5 +3,5 @@ package com.nehvedovich.vladimir.pogoda.screens.rest.entites;
 import com.google.gson.annotations.SerializedName;
 
 public class CloudsRestModel {
-    @SerializedName("all") public int all;
+    @SerializedName("all") public float all;
 }
